+++ cognite/neat/_rules/models/dms/_rules_input.py
@@ -5,6 +5,7 @@ from typing import Any, Literal
 
 import pandas as pd
 from cognite.client import data_modeling as dm
+from cognite.client.data_classes.data_modeling import ContainerId, ViewId
 from rdflib import Namespace, URIRef
 
 from cognite.neat._constants import DEFAULT_NAMESPACE
@@ -125,6 +126,12 @@ class DMSInputProperty(InputComponent[DMSProperty]):
         )
         return output
 
+    def referenced_view(self, default_space: str, default_version: str) -> ViewEntity:
+        return ViewEntity.load(self.view, strict=True, space=default_space, version=default_version)
+
+    def referenced_container(self, default_space: str) -> ContainerEntity | None:
+        return ContainerEntity.load(self.container, strict=True, space=default_space) if self.container else None
+
 
 @dataclass
 class DMSInputContainer(InputComponent[DMSContainer]):
@@ -140,14 +147,16 @@ class DMSInputContainer(InputComponent[DMSContainer]):
 
     def dump(self, default_space: str) -> dict[str, Any]:  # type: ignore[override]
         output = super().dump()
-        container = ContainerEntity.load(self.container, space=default_space)
-        output["Container"] = container
+        output["Container"] = self.as_entity_id(default_space)
         output["Constraint"] = (
             [ContainerEntity.load(constraint.strip(), space=default_space) for constraint in self.constraint.split(",")]
             if self.constraint
             else None
         )
         return output
+
+    def as_entity_id(self, default_space: str) -> ContainerEntity:
+        return ContainerEntity.load(self.container, strict=True, space=default_space)
 
     @classmethod
     def from_container(cls, container: dm.ContainerApply) -> "DMSInputContainer":
@@ -182,17 +191,25 @@ class DMSInputView(InputComponent[DMSView]):
 
     def dump(self, default_space: str, default_version: str) -> dict[str, Any]:  # type: ignore[override]
         output = super().dump()
-        view = ViewEntity.load(self.view, space=default_space, version=default_version)
-        output["View"] = view
-        output["Implements"] = (
+        output["View"] = self.as_entity_id(default_space, default_version)
+        output["Implements"] = self._load_implements(default_space, default_version)
+        return output
+
+    def as_entity_id(self, default_space: str, default_version: str) -> ViewEntity:
+        return ViewEntity.load(self.view, strict=True, space=default_space, version=default_version)
+
+    def _load_implements(self, default_space: str, default_version: str) -> list[ViewEntity] | None:
+        return (
             [
-                ViewEntity.load(implement, space=default_space, version=default_version)
+                ViewEntity.load(implement, strict=True, space=default_space, version=default_version)
                 for implement in self.implements.split(",")
             ]
             if self.implements
             else None
         )
-        return output
+
+    def referenced_views(self, default_space: str, default_version: str) -> list[ViewEntity]:
+        return self._load_implements(default_space, default_version) or []
 
     @classmethod
     def from_view(cls, view: dm.ViewApply, in_model: bool) -> "DMSInputView":
@@ -287,3 +304,30 @@ class DMSInputRules(InputRules[DMSRules]):
         return DEFAULT_NAMESPACE[
             f"data-model/unverified/dms/{self.metadata.space}/{self.metadata.external_id}/{self.metadata.version}"
         ]
+
+    def referenced_views_and_containers(self) -> tuple[set[ViewEntity], set[ContainerEntity]]:
+        default_space = self.metadata.space
+        default_version = self.metadata.version
+
+        containers: set[ContainerEntity] = set()
+        views = {parent for view in self.views for parent in view.referenced_views(default_space, default_version)}
+        for prop in self.properties:
+            views.add(prop.referenced_view(default_space, default_version))
+            if ref_container := prop.referenced_container(default_space):
+                containers.add(ref_container)
+
+        return views, containers
+
+    def as_view_entities(self) -> list[ViewEntity]:
+        return [view.as_entity_id(self.metadata.space, self.metadata.external_id) for view in self.views]
+
+    def as_container_entities(self) -> list[ContainerEntity]:
+        return [container.as_entity_id(self.metadata.space) for container in self.containers or []]
+
+    def imported_views_and_containers(self) -> tuple[set[ViewEntity], set[ContainerEntity]]:
+        views, containers = self.referenced_views_and_containers()
+        return views - set(self.as_view_entities()), containers - set(self.as_container_entities())
+
+    def imported_views_and_containers_ids(self) -> tuple[set[ViewId], set[ContainerId]]:
+        views, containers = self.imported_views_and_containers()
+        return {view.as_id() for view in views}, {container.as_id() for container in containers}